Move db models into their own files
Also rename Team database model to TeamComposition so as not to conflict with Team class.
