+++ src/db/create.py
@@ -1,3 +1,5 @@
 from src.web.server import db
+from src.db.models.pick import Pick
+from src.db.models.team_composition import TeamComposition
 
 db.create_all()
